ViewFlow - Tweaks to touch handling
1. Removed code from onInterceptTouchEvent and simply return true.
Returning true ensures this ViewGroup intercepts the MotionEvent however
no need to repeat code from onTouchEvent as that method is called
automatically after onInterceptTouchEvent.

2. Added additional else if (mTouchState == TOUCH_STATE_REST) to
ACTION_UP event in onTouchEvent to essentially detect a 'tap/click' as
opposed to a 'drag'. This calls performItemClick() which calls
onItemClickListener.onItemClick(). If there is no onItemClickListener
this addition has no effect.

3. Added a field for parentFriendlyScrollView and logic to disable the
vertical scrolling of the FriendlyScrollView if the ViewFlow is
scrolling. If no FriendlyScrollView is supplied then these changes have
no effect.